% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{dsgd_simple}
\alias{dsgd_simple}
\title{Disease-Specific Gene Detection in Simple Scenario}
\usage{
dsgd_simple(list_y, matrix_x, num_samples, num_covariates)
}
\arguments{
\item{list_y}{a list of binary variables to indicate the spot status}

\item{matrix_x}{a martix of covaraites (Important: coordinates must be placed at the last columns of matrix)}

\item{num_samples}{number of spots}

\item{num_covariates}{number of genes}
}
\value{
a fitted stan model
}
\description{
Disease-Specific Gene Detection in Simple Scenario
}
\examples{
\donttest{}
}
